config/requirements.txt
channels==4.0.0
Django==4.2.5
Django==5.0.2
djangorestframework==3.14.0
djoser==2.2.2
